Quick cleanup
